requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
dashscope>=1.20.0
faiss-cpu>=1.8.0
python-multipart>=0.0.9
pydantic>=2.0.0
pymupdf>=1.24.0
python-docx>=1.0.0
pyyaml>=6.0
